refactored code into three subpackages
