+++ app/src/main/java/Main.java
@@ -13,6 +13,7 @@ import java.util.Map;
 
 import com.mu_bball_stats.database.*;
 import com.mu_bball_stats.web.*;
+import com.mu_bball_stats.model.*;
 import com.mu_bball_stats.WebBrowser;
 
 public class Main {
