+++ app/src/main/java/com/mu_bball_stats/database/DBTableManager.java
@@ -9,6 +9,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.TreeMap;
 
+import com.mu_bball_stats.model.Player;
+import com.mu_bball_stats.model.PlayerStat;
+import com.mu_bball_stats.model.Roster;
+
 /**
  * Implements the RosterDataManager interface and provides methods
  * for managing player data stored in the Players database table.
