+++ app/src/main/java/com/mu_bball_stats/model/SessionStat.java
@@ -1,4 +1,4 @@
-package com.mu_bball_stats.database;
+package com.mu_bball_stats.model;
 
 import java.util.Map;
 import java.util.HashMap;
