+++ app/src/main/jte/roster.jte
@@ -1,5 +1,5 @@
-@import com.mu_bball_stats.database.Roster
-@import com.mu_bball_stats.database.Player
+@import com.mu_bball_stats.model.Roster
+@import com.mu_bball_stats.model.Player
 @import com.mu_bball_stats.web.Page
 
 @param Page page
